pyproject.toml
[project]
name = "datasets-dump"
version = "0.1.1"
description = "A tool for dumping datasets from the Hugging Face datasets library"
authors = [{ name = "Jacob Lim", email = "jacoblincool@gmail.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["datasets"]
dependencies = ["datasets", "tqdm", "Pillow", "soundfile", "numpy"]

[project.urls]
homepage = "https://github.com/JacobLinCool/datasets-dump"
repository = "https://github.com/JacobLinCool/datasets-dump"

[project.scripts]
datasets-dump = "datasets_dump.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["datasets_dump/*.py"]
